# src/api-doc.json
{
    "openapi":"3.0.0",
    "info": {
      "title":"Swagger Kjaga Backend",
      "version":"1.0",
      "description": "Documentation for Kjaga backend",
      "contact": {
        "name": "Dimas Faidh Muzaki",
        "email": "dimasfaid@gmail.com"
      }
    },
    "servers": [
      {
        "url":"http://localhost:3001",
        "description": "local"
      },
      {
        "url": "https://capstone-bangkit-kjaga.et.r.appspot.com",
        "description": "App engine"
      }
    ],
    "externalDocs": {
      "url": "https://github.com/maspaitujaki/backend-capstone-bangkit",
      "description": "Private Repository"
    },
    "tags": [
      {
      "name": "User",
      "description": "Operations about user"
      },
      {
        "name": "Food",
        "description": "Operations about food and portions"
      }
    ],
    "paths": {
      "/users": {
        "post": {
          "operationId": "createUser",
          "tags": ["User"],
          "description": "Use as register",
          "summary": "Create a user",
          "requestBody" : {
            "required": true,
            "content": {
              "application/json": {
                "schema" : {
                  "type": "object",
                  "properties": {
                    "email" : {
                      "type": "string",
                      "format": "email"
                    },
                    "username" : {
                      "type": "string"
                    },
                    "password" : {
                      "type": "string"
                    },
                    "confirmPassword" : {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201" : {
              "description": "User successfully created",
              "content": {
                "application/json": {
                  "schema" : {
                    "type": "object",
                    "properties": {
                      "id" : {
                        "type": "string"
                      },
                      "email" : {
                        "type": "string",
                        "format": "email"
                      },
                      "username" : {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              } 
            },
            "400" : {
              "description": "Username or email already registered",
              "content": {
                "application/json": {
                  "schema":{
                    "$ref": "#/components/schemas/errorMessage"
                  }
                }
              }
            }
          }
        }
      },
      "/users/login": {
        "post": {
          "operationId": "loginUser",
          "tags": ["User"],
          "description": "Use as sign in",
          "summary": "Logs user into the system",
          "requestBody":{
            "required": true,
            "content": {
              "application/json":{
                "schema":{
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password":{
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200" : {
              "description": "Successfully logged in",
              "content": {
                "application/json": {
                  "schema" : {
                    "type": "object",
                    "properties": {
                      "user": {
                        "type": "object",
                        "properties": {
                          "id" : {
                            "type": "string"
                          },
                          "email" : {
                            "type": "string",
                            "format": "email"
                          },
                          "username" : {
                            "type": "string"
                          }
                        }
                      },
                      "token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401" : {
              "description": "Combination of username and password is not found",
              "content": {
                "application/json": {
                  "schema":{
                    "$ref": "#/components/schemas/errorMessage"
                  }
                }
              }
            }
          }
        }
      },
      "/users/{username}":{
        "get": {
          "operationId": "readUserByUsername",
          "tags": ["User"],
          "description": "",
          "summary": "Get user by username",
          "parameters": [{
            "$ref": "#/components/parameters/userUsername"
          }],
          "responses": {
            "200": {
              "description": "",
              "content": {
                "application/json":{
                  "schema":{
                    "$ref": "#/components/schemas/user"
                  }
                }
              }
            },
            "400": {
              "description": "Username not exist",
              "content": {
                "application/json":{
                  "schema":{
                    "$ref": "#/components/schemas/errorMessage"
                  }
                }
              }
            }
          }     
        },
        "put": {
          "operationId": "updateUserByUsername",
          "tags": ["User"],
          "description": "",
          "summary": "Update user by username, certain field only",
          "parameters": [{
            "$ref": "#/components/parameters/userUsername"
          }],
          "requestBody":{
            "required": true,
            "content": {
              "application/json":{
                "schema":{
                  "type": "object",
                  "properties": {
                    "email":{
                      "type": "string",
                      "format": "email"
                    },
                    "username":{
                      "type": "string"
                    },
                    "gender": {
                      "type": "string",
                      "enum": ["male","female"]
                    },
                    "birthdate":{
                      "type": "string",
                      "format": "date"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "",
              "content": {
                "application/json":{
                  "schema":{
                    "$ref": "#/components/schemas/user"
                  }
                }
              }
            },
            "400": {
              "description": "Username not exist",
              "content": {
                "application/json":{
                  "schema":{
                    "$ref": "#/components/schemas/errorMessage"
                  }
                }
              }
            }
          }        
        },
        "delete": {
          "operationId": "deleteUserByUsername",
          "tags": ["User"],
          "description": "",
          "summary": "Delete user by username",
          "parameters": [{
            "$ref": "#/components/parameters/userUsername"
          }],
          "responses": {
            "200":{
              "description": "Successfully deleted",
              "content": {
                "application/json":{
                  "schema":{
                    "type": "object",
                    "properties": {
                      "message":{
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Username not exist",
              "content": {
                "application/json":{
                  "schema":{
                    "$ref": "#/components/schemas/errorMessage"
                  }
                }
              }
            }
          }       
        } 
      },
      "/foods":{
        "get":{
          "operationId": "getFoods",
          "description": "Search foods",
          "tags": ["Food"],
          "parameters": [
            {
              "schema" : {
                "type": "string"
              },
              "in": "query",
              "name": "search",
              "description": "Search food by name"
            }
          ],
          "responses": {
            "200": {
              "description": "Return a list of foods according to search query. If query is empty then return all. . Note that the portion is default portion",
              "content": {
                "application/json":{
                  "schema" :{
                    "type": "object",
                    "properties": {
                      "count": { 
                        "type": "integer"
                      },
                      "foods" : {
                        "type": "array",
                        "items" : {
                          "type": "object",
                          "properties": {
                            "food": {
                              "$ref": "#/components/schemas/food"
                            },
                            "portion_count":{
                              "description": "How many available portion for this food",
                              "type": "integer"
                            }, 
                            "portion":{
                              "$ref": "#/components/schemas/portion"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/foods/{food_id}/portions": {
        "get":{
          "operationId": "getPortionByFoodId",
          "tags": ["Food"],
          "parameters": [
            {
              "name": "food_id",
              "description": "The food id",
              "in": "path",
              "required": true,
              "schema" : {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200":{
              "description": "Return the portion that available for a food",
              "content": {
                "application/json":{
                  "schema":{
                    "type": "object",
                    "properties": {
                      "food":{
                        "$ref": "#/components/schemas/food"
                      },
                      "portions":{
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/portion"
                        }
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Food not exist",
              "content": {
                "application/json":{
                  "schema":{
                    "$ref": "#/components/schemas/errorMessage"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "parameters": {
        "userUsername": {
          "schema":{
            "type": "string"
          },
          "in": "path",
          "name": "username",
          "description": "Username to be read",
          "required": true
        }
      },
      "schemas": {
        "errorMessage": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string"
            }
          }
        },
        "user":{
          "type": "object",
          "properties": {
            "id" : {
              "type": "string"
            },
            "email" : {
              "type": "string",
              "format": "email"
            },
            "username" : {
              "type": "string"
            },
            "gender": {
              "type": "string",
              "enum": ["pria","wanita"]
            },
            "birthdate":{
              "type": "string",
              "format": "date"
            },
            "age":{
              "type": "integer"
            },
            "akg_type":{
              "type": "string"
            },
            "created_at": {
              "type": "string",
              "format": "date-time"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "nutrition": {
          "type": "object",
          "properties": {
            "energy(kkal)" : {
              "type": "number"
            },
            "fat(g)" : {
              "type": "number"
            },
            "cholesterol(mg)" : {
              "type": "number"
            },
            "protein(g)" : {
              "type": "number"
            },
            "carbohydrates(g)" : {
              "type": "number"
            },
            "fiber(g)" : {
              "type": "number"
            },
            "sugar(g)" : {
              "type": "number"
            },
            "sodium(mg)" : {
              "type": "number"
            },
            "kalium(mg)" : {
              "type": "number"
            }
          }
        },
        "portion":{
          "type": "object",
          "properties": {
            "id":{
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "nutrition":{
              "$ref": "#/components/schemas/nutrition"
            }
          }
        },
        "food":{
          "type": "object",
          "properties": {
            "id":{
              "type": "string"
            },
            "name":{
              "type": "string"
            }
          }
        }
      }
    }
  }